+++ core/src/bricks/BrickTypes.cpp
@@ -6,13 +6,10 @@
 
 BrickType BrickTypeFromString(const string& typeStr) {
     static const std::unordered_map<string, BrickType> typeMap = {
-        {"storage", BrickType::Storage},
-        {"generic_land_cover", BrickType::GenericLandCover},
-        {"ground", BrickType::GenericLandCover},
-        {"generic", BrickType::GenericLandCover},
-        {"glacier", BrickType::Glacier},
-        {"snowpack", BrickType::Snowpack},
-        {"interception_storage", BrickType::InterceptionStorage}};
+        {"storage", BrickType::Storage},         {"generic_land_cover", BrickType::GenericLandCover},
+        {"ground", BrickType::GenericLandCover}, {"generic", BrickType::GenericLandCover},
+        {"open", BrickType::GenericLandCover},   {"glacier", BrickType::Glacier},
+        {"snowpack", BrickType::Snowpack},       {"interception_storage", BrickType::InterceptionStorage}};
 
     auto it = typeMap.find(typeStr);
     if (it != typeMap.end()) {
@@ -22,7 +19,7 @@ BrickType BrickTypeFromString(const string& typeStr) {
 }
 
 vector<string> GetValidBrickTypes() {
-    static const vector<string> validTypes = {"storage",  "generic_land_cover",  "ground", "generic", "glacier",
+    static const vector<string> validTypes = {"storage",  "generic_land_cover",  "ground", "generic", "open", "glacier",
                                               "snowpack", "interception_storage"};
     return validTypes;
 }